updated readme, added dmg bundling for macos targets, added win7 target option

+++ build-releases.sh
@@ -12,6 +12,11 @@ else
   exit 2
 fi
 
+if ! command -v mkfs.hfsplus >/dev/null 2>/dev/null; then
+  echo "mkfs.hfsplus not found on your path"
+  exit 3
+fi
+
 # install latest pkg version
 echo installing latest version of pkg bundler
 sudo npm install --global pkg 
@@ -19,9 +24,7 @@ echo '------------------------------------------------------'
 echo 
 
 # check if pkg binary is available
-if pkg --help >/dev/null 2>&1; then
-  sleep 0
-else
+if ! command -v pkg >/dev/null 2>&1; then
   echo "pkg bundler not found on your PATH."
   exit 1
 fi 
@@ -35,26 +38,41 @@ if [[ "$sampleManVersion" == "" ]]; then
 else
   echo "BUILDING SAMPLEMAN $sampleManVersion RELEASES"
 fi
-echo "node --version is v$(node --version | grep -oP '\d+\.\d+\.\d+')"
+echo "> node --version is v$(node --version | grep -oP '\d+\.\d+\.\d+')"
 echo
 
 # build targets 
-for target in "linux-x64" "win-x64" "macos-x64" "macos-arm64"; do
+#for target in "linux-x64" "win10-x64" "win7-x64" "macos-x64" "macos-arm64"; do
+for target in "linux-x64" "win10-x64" "macos-x64" "macos-arm64"; do
 
   echo "begin building target $target..."
 
-  # check latest nodejs prebuilt binary version
-  npm install --no-save semver-sort >/dev/null 2>&1
-  echo "const s = require('semver-sort'); \
-        console.log(s.desc(process.argv.slice(2))[0])" >.tmp-script.js
-  latest="$(node .tmp-script.js $(curl "https://github.com/vercel/pkg-fetch/releases" 2>/dev/null \
-              | grep -oP '\d+\.\d+\.\d+-'"$target" | grep -oP '\d+\.\d+\.\d+'))"
-  rm .tmp-script.js
-  echo "> using prebuilt node v$latest"
+  # mangling win7 and win10 targets
+  effTarget="$target"
+  nodeVersion=""
+  if [[ "$target" == "win7-x64" ]]; then
+    effTarget="win-x64"
+    nodeVersion="12.22.2"
+  fi
+  if [[ "$target" == "win10-x64" ]]; then
+    effTarget="win-x64"
+  fi
+
+  # check latest nodejs prebuilt binary version if none is set yet
+  if [[ "$nodeVersion" == "" ]]; then
+    npm install --no-save semver-sort >/dev/null 2>&1
+    echo "const s = require('semver-sort'); \
+          console.log(s.desc(process.argv.slice(2))[0])" >.tmp-script.js
+    nodeVersion="$(node .tmp-script.js $(curl "https://github.com/vercel/pkg-fetch/releases" 2>/dev/null \
+                    | grep -oP '\d+\.\d+\.\d+-'"$effTarget" | grep -oP '\d+\.\d+\.\d+'))"
+    rm .tmp-script.js
+  fi
+
+  echo "> using prebuilt node v$nodeVersion"
 
   mkdir -p releases/
-  outname="sampleman-$sampleManVersion-$target-node$latest"
-  pkg . --targets "node$latest-$target" \
+  outname="sampleman-$sampleManVersion-$target-node$nodeVersion"
+  pkg . --targets "node$nodeVersion-$effTarget" \
         --output "releases/$outname" \
           | grep -i "error!"
   if [[ "${PIPESTATUS[0]}" == "0" ]]; then
@@ -65,6 +83,30 @@ for target in "linux-x64" "win-x64" "macos-x64" "macos-arm64"; do
   fi
   echo
 done
+
+# store macos build in dmg containers
+echo "stuffing macos builds into .dmg containers..."
+echo
+
+mkdir -p "mnt"
+for f in $(ls "releases"); do
+  if [[ "$(echo "$f" | cut -d'-' -f3)" == "macos" ]] && [[ "${f: -3}" != "dmg" ]]; then
+    echo "begin packing $f..."
+    dd if=/dev/zero of="releases/$f.dmg" bs=1M count=$(expr $(du -BM "releases/$f" | grep -oP '^\d+') + 3) >/dev/null 2>&1 \
+      && mkfs.hfsplus -v sampleman "releases/$f.dmg" >/dev/null \
+      && sudo mount -o loop "releases/$f.dmg" "mnt" \
+      && sudo cp -a "releases/$f" "mnt/" \
+      && sudo umount "mnt" \
+      && rm "releases/$f"
+    if [[ -e "releases/$f" ]]; then
+      echo "> failed."
+    else
+      echo "> done."
+    fi
+    echo
+  fi
+done
+rmdir "mnt"
 
 # report how many binaries were created
 if [[ "$created" == "" ]]; then
